+++ src/javascript/binary/websocket_pages/user/account/change_password.js
@@ -1,5 +1,8 @@
 var testPassword = require('../../../common_functions/passwordmeter').testPassword;
 var ValidateV2 = require('../../../common_functions/validation').ValidateV2;
+var ValidationUI = require('../../../validator').ValidationUI;
+var customError = require('../../../validator').customError;
+var bind_validation = require('../../../validator').bind_validation;
 
 var PasswordWS = (function(){
     var $form, $result;
