+++ src/javascript/binary/websocket_pages/user/account/settings/api_token.js
@@ -2,6 +2,9 @@ var showLoadingImage = require('../../../../base/utility').showLoadingImage;
 var showLocalTimeOnHover = require('../../../../base/utility').showLocalTimeOnHover;
 var FlexTableUI = require('../../../../common_functions/attach_dom/flextable').FlexTableUI;
 var ValidateV2 = require('../../../../common_functions/validation').ValidateV2;
+var ValidationUI = require('../../../../validator').ValidationUI;
+var customError = require('../../../../validator').customError;
+var bind_validation = require('../../../../validator').bind_validation;
 
 var APITokenWS = (function() {
     "use strict";
