+++ src/javascript/binary/websocket_pages/user/account/settings/settings_detailsws.js
@@ -1,4 +1,5 @@
 var ValidateV2 = require('../../../../common_functions/validation').ValidateV2;
+var bind_validation = require('../../../../common_functions/validation').bind_validation;
 
 var SettingsDetailsWS = (function() {
     "use strict";
