+++ src/javascript/binary/websocket_pages/user/new_account/virtual_acc_opening.js
@@ -1,5 +1,6 @@
 var template = require('../../../base/utility').template;
 var testPassword = require('../../../common_functions/passwordmeter').testPassword;
+var bind_validation = require('../../../validator').bind_validation;
 
 pjax_config_page("new_account/virtualws", function() {
     function onSuccess(res) {
